refactor: extract weapon roll orchestration to weapon-rolls.mjs
- Add module/sheets/actor/weapon-rolls.mjs (attack dialog, processWeaponRoll, ROF rerolls)
- Extend roll-math.mjs with success die, skill dice, melee STR damage, explosive helpers
- Add unit tests for new roll-math pure functions (67 total)
- Delegate from SlaActorSheet; shrink actor-sheet by ~400 lines
